Start of implementation of desktop Anki's SRS
Mostly a straight port of the classes and methods from the python code.
Code should be in place for getting the next scheduled card, but this is
not used to get the new card currently in the UI (ie. it is not tested
yet).

+++ src/com/ichi2/anki/AnkiDb.java
@@ -47,12 +47,12 @@ public class AnkiDb {
 	 * @param query The raw SQL query to use.
 	 * @return The integer result of the query.
 	 */
-	static public int queryScalar(String query) throws SQLException {
+	static public long queryScalar(String query) throws SQLException {
 		Cursor cursor = AnkiDb.database.rawQuery(query, null);
 		if (!cursor.moveToFirst())
 			throw new SQLException("No result for query: " + query);
 		
-		int scalar = cursor.getInt(0);
+		long scalar = cursor.getLong(0);
 		cursor.close();
 		
 		return scalar;
